Move MovieCode inside Movie record
